pyproject.toml
[tool.poetry]
name = "juniper-srx-security-demo"
version = "0.1.0"
description = "Python packages to automate the deployment of firewall"
authors = ["Calvin Remsburg <cremsburg.dev@gmail.com>"]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = "^3.8"
junos-eznc = "^2.6.3"
jsnapy = "^1.3.6"
Jinja2 = "^3.0.3"
invoke = "^1.6.0"
toml = "^0.10.2"
jxmlease = "^1.0.3"
types-PyYAML = "^6.0.3"
pre-commit = "^2.17.0"
icdiff = "1.9.1"
pydantic = "^1.9.1"

[tool.poetry.dev-dependencies]
black = {version = "^21.12b0", allow-prereleases = true}
isort = "^5.10.1"
autoflake = "^1.4"
bandit = "^1.7.1"
flake8 = "^4.0.1"
flake8-print = "^4.0.0"
flake8-builtins = "^1.5.3"
flake8-eradicate = "^1.2.0"
pydocstyle = "^6.1.1"
yamllint = "^1.26.3"
mypy = "^0.931"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
